requirements.txt
asgiref==3.11.0
charset-normalizer==3.4.4
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
pillow==12.1.0
psycopg2-binary==2.9.11
python-barcode==0.16.1
python-decouple==3.8
reportlab==4.4.9
sqlparse==0.5.5
tzdata==2025.3
